Remove unused setup
This file may have been misnamed. The comments suggest that it's for
compatibility with `rvm` or `avn`, and I'm not using either of those.
Without the file, using a `.env` file in a Python project still works
fine.

I'm assuming it's safe to get rid of and that since `rvm` has likely
changed a lot, it may not apply to that tool any more, and `avn` is no
longer actively worked on.